Finished section1-content & Edited Nav-branch
Content was added to the first section of the landing
page. The first section consists of a welcome message,
submessage, sign-up button and an image. The nav-bar
was also edited to make the nav-bar items have an
underline once hovered. The nav bar section was also
cleaned up and un-used properties were removed.